llvm-kernel-testing: Add initial testing framework
We're kind of bad at testing LLVM releases because we focus so heavily
on making sure that tip of tree stays green. The idea behind this
framework is to build a revision of LLVM that we supply (by default, the
latest stable release or release candidate), build every single config
we know builds on the kernel version we are testing against, and boot
them if necessary (defconfigs).

This is very similar to our current continuous integration setup but it
differs in the sense that we want to build ALL known working configs,
not just defconfigs minus modules. This could in theory be wired up to
some continuous integration setup since it is just a self contained
script; however, this does assume that computing resources are not
necessarily a concern because this builds several large kernels.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>