lkt: Deduplicate CONFIG_WERROR checks for allmodconfig
We never want -Werror for allmodconfig and LinuxSourceManager is now
required for LLVMRunner, so move adding CONFIG_WERROR=n to the
configuration list into LLVMRunner when allmodconfig is in the list of
targets. This not only cleans up code but it is less error prone every
time that we add support for a new architecture (which happens more
often than you might think...).

Signed-off-by: Nathan Chancellor <nathan@kernel.org>