lkt: powerpc: Adjust logic for enabling allmodconfig build
The stable team backported a11334d8327b3f without its dependencies,
which breaks the assumptions that I had when enabling this build about
having all dependencies if the contents of that change were present.
Just switch to a version check, as there are a number of changes needed
for a working allmodconfig build in 6.1 that are probably not worth
requesting given how simple this fix is.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>